all: a merger between two implementations
This is a merger between eric-ch/openxt-manifest and
apertussolutions/openxt-manifest. It takes erich-ch's openxt.sh and set-env
and breaks up into a command module solution that is merged with
apertussolutions' set of scripts. As a result, there is a bin and templates
directories that houses the system. Under bin is the bordel script that
builds up its list of subcommands from command modules under bin/cmds and a
common function library. The README provides a more detailed explanation of
how the new solution works.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>